+++ tests/extract_resources.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from test_res import resources
+
+resources.extract("resource_extracted", force = False)